Flattened ApiKeyInfoResponse.
