+++ src/main/java/com/beimin/eveapi/utils/AccessChecker.java
@@ -3,12 +3,12 @@ package com.beimin.eveapi.utils;
 import java.util.HashSet;
 import java.util.Set;
 
-import com.beimin.eveapi.model.account.ApiKeyInfo;
 import com.beimin.eveapi.model.calllist.Call;
 import com.beimin.eveapi.model.calllist.CallList;
+import com.beimin.eveapi.response.account.ApiKeyInfoResponse;
 
 public class AccessChecker {
-    public static Set<Call> getCalls(final ApiKeyInfo apiKeyInfo, final CallList callList) {
+    public static Set<Call> getCalls(final ApiKeyInfoResponse apiKeyInfo, final CallList callList) {
         final Set<Call> result = new HashSet<Call>();
         for (final Call call : callList.getCalls()) {
             final long cam = call.getAccessMask();
