+++ src/test/java/com/beimin/eveapi/account/apikeyinfo/ApiKeyInfoParserOnlineTest.java
@@ -3,7 +3,6 @@ package com.beimin.eveapi.account.apikeyinfo;
 import org.junit.Test;
 
 import com.beimin.eveapi.AbstractOnlineTest;
-import com.beimin.eveapi.model.account.ApiKeyInfo;
 import com.beimin.eveapi.parser.account.ApiKeyInfoParser;
 import com.beimin.eveapi.response.account.ApiKeyInfoResponse;
 
@@ -12,7 +11,7 @@ public class ApiKeyInfoParserOnlineTest extends AbstractOnlineTest {
     @Test
     public void getResponse() throws Exception {
         setAlias(com.beimin.eveapi.model.account.Character.class, "characterName", "name");
-        setAlias(ApiKeyInfo.class, "characters", "eveCharacters");
+        setAlias(ApiKeyInfoResponse.class, "characters", "eveCharacters");
         ignoreXmlField("key");
         // Not in alliance
         allowEmpty("getAllianceName");
